Break out lifted functions
